project/requirements.txt
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
Django==4.2.23
idna==3.10
importlib_metadata==8.7.0
internetarchive==5.4.0
jsonpatch==1.33
jsonpointer==3.0.0
python-dateutil==2.9.0.post0
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.14.0
urllib3==2.5.0
zipp==3.23.0
